refactor: Fix bug in equipamentosController.js
This commit fixes a bug in the equipamentosController.js file where the condition for checking if the equipment already exists was incorrect. The bug caused the wrong error message to be returned. The issue has been resolved by changing the condition to check if the equipment exists with the message 'Patrimonio já existe'.